Rename: 폴더 정리 및 파일 이름 수정
