+++ src/main/java/com/ourhours/server/global/model/exception/JwtException.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class JwtException extends BaseException {
-
 	public JwtException(ExceptionConstant exceptionConstant) {
 		super(exceptionConstant);
 	}
