+++ src/main/java/com/ourhours/server/global/model/jwt/dto/request/JwtAuthenticationRequestDto.java
@@ -1,4 +1,4 @@
-package com.ourhours.server.global.model.security.dto.request;
+package com.ourhours.server.global.model.jwt.dto.request;
 
 public record JwtAuthenticationRequestDto(String token, Long userId) {
 }
