+++ src/main/java/com/ourhours/server/global/model/security/JwtAuthentication.java
@@ -7,7 +7,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
-import com.ourhours.server.global.model.security.dto.request.JwtAuthenticationRequestDto;
+import com.ourhours.server.global.model.jwt.dto.request.JwtAuthenticationRequestDto;
 
 import lombok.Getter;
 
@@ -16,18 +16,15 @@ public class JwtAuthentication implements Authentication {
 
 	private final String token;
 	private final Long userId;
-	private boolean isAuthenticated;
 
 	public JwtAuthentication(JwtAuthenticationRequestDto dto) {
 		this.token = dto.token();
 		this.userId = dto.userId();
-		this.isAuthenticated = true;
 	}
 
 	@Override
 	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return isAuthenticated ? Collections.singletonList(new SimpleGrantedAuthority(Role.USER.name())) :
-			Collections.singletonList(new SimpleGrantedAuthority(Role.ANONYMOUS.name()));
+		return Collections.singletonList(new SimpleGrantedAuthority(Role.USER.name()));
 	}
 
 	@Override
@@ -47,12 +44,11 @@ public class JwtAuthentication implements Authentication {
 
 	@Override
 	public boolean isAuthenticated() {
-		return this.isAuthenticated;
+		return true;
 	}
 
 	@Override
 	public void setAuthenticated(boolean isAuthenticated) throws IllegalArgumentException {
-		this.isAuthenticated = isAuthenticated;
 
 	}
