+++ src/main/java/com/ourhours/server/global/util/jwt/filter/JwtExceptionHandlerFilter.java
@@ -28,7 +28,7 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Component
-public class JwtExceptionFilter extends OncePerRequestFilter {
+public class JwtExceptionHandlerFilter extends OncePerRequestFilter {
 	@Override
 	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
 		FilterChain filterChain) throws ServletException, IOException {
